+++ controller/main.go
@@ -20,7 +20,7 @@ var assets embed.FS
 
 // Version information - injected at build time via -ldflags -X
 var (
-	Version     = "v3.1.88"
+	Version     = "v3.1.89"
 	BuildDate   = "2026-05-31"
 	VersionInfo = ""
 )